Add. parser.js 모듈 생성, renderRecommendedKeywords 함수를 rolling.js 모듈로 이동
