+++ public/js/util/parser.js
@@ -1,0 +1,6 @@
+// json 데이터에서 필요한 데이터만 뽑는 함수들
+
+// 인기 검색어 1~10위를 담은 배열
+const recommendParser = (data) => data.map((v) => v.keyword).slice(0, 10);
+
+export { recommendParser }
